PET-211 Refactor: MemberFactory 패키지 변경
